--- pubnub-core/src/subscribe/mod.rs
@@ -1,10 +1,0 @@
-pub(crate) mod channel;
-
-mod encoded_channels_list;
-mod registry;
-mod subscribe_loop;
-mod subscribe_request;
-mod subscription;
-
-pub use subscribe_loop::*;
-pub use subscription::*;
