+++ pubnub-core/src/subscription/registry.rs
